ima: Move to LSM infrastructure
Remove hardcoded IMA function calls (not for appraisal) from the LSM
infrastructure or from the other places that call IMA functions directly.

Make those functions as static (except for ima_file_check() which is
exported), and register them as implementation of the respective hooks in
the new function init_ima_lsm().

Call init_ima_lsm() from the integrity code, as there will be one
'integrity' LSM. The same will be done for EVM, by calling init_evm_lsm()
just after init_ima_lsm().

Introducing the 'integrity' LSM instead of two separate LSMs for IMA and
EVM is the safest solution, to avoid for example to accidentally allow file
writes without updating the HMAC, which would happen if EVM is not an
enabled LSM.

Systems updating the HMAC currently need the 'integrity' LSM. Without it,
the kernel would panic. By keeping the same name, this ensures that after
the transition from an old kernel to a new kernel IMA and EVM hooks are
executed as before.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>