security: Restore passing final prot to ima_file_mmap()
Commit 98de59bfe4b2f ("take calculation of final prot in
security_mmap_file() into a helper") moved the code to update prot with the
actual protection flags granted to the requestor by the kernel to a helper
called mmap_prot(). However, the patch didn't update the argument passed to
ima_file_mmap(), making it receive the requested prot instead of the final
computed prot.

Replace prot with mmap_prot(prot) as second argument to ima_file_mmap() to
restore the original behavior.

Cc: stable@vger.kernel.org
Fixes: 98de59bfe4b2 ("take calculation of final prot in security_mmap_file() into a helper")
Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>