security: Introduce LSM_ORDER_LAST
Introduce LSM_ORDER_LAST, to satisfy the requirement of LSMs willing to be
the last, e.g. the 'integrity' LSM, without changing the kernel command
line or configuration.

If there are multiple LSMs with order set to LSM_ORDER_LAST, their relative
order will be respected, as for LSM_ORDER_FIRST.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>