ima: Move to LSM infrastructure
Remove hardcoded IMA function calls (not for appraisal) from the LSM
infrastructure, the VFS, NFS and the key subsystem.

Make those functions as static (except for ima_file_check() which is
exported, and ima_post_key_create_or_update(), which is not in ima_main.c),
and register them as implementation of the respective hooks in the new
function init_ima_lsm().

Call init_ima_lsm() from integrity_lsm_init() (renamed from
integrity_iintcache_init()), to make sure that the integrity subsystem is
ready at the time IMA hooks are registered. The same will be done for EVM,
by calling init_evm_lsm() just after init_ima_lsm().

Put 'integrity' at the end of the LSM list by setting the order to
LSM_ORDER_LAST. 'integrity' must be the last so that EVM can calculate the
HMAC on xattrs previously provided by the other selected LSMs.

Setting the order to LSM_ORDER_LAST also causes 'integrity' to be always
enabled regardless of the LSM choice. This avoids any risk that IMA and EVM
functions are not called when transitioning from an old kernel to one where
IMA and EVM are stacked.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>