ima: Move to LSM infrastructure
Remove hardcoded IMA function calls (not for appraisal) from the LSM
infrastructure, the VFS, NFS and the key subsystem.

Make those functions as static (except for ima_file_check() which is
exported, and ima_post_key_create_or_update(), which is not in ima_main.c),
and register them as implementation of the respective hooks in the new
function init_ima_lsm().

Call init_ima_lsm() from integrity_lsm_init() (renamed from
integrity_iintcache_init()), to make sure that the integrity subsystem is
ready at the time IMA hooks are registered. The same will be done for EVM,
by calling init_evm_lsm() just after init_ima_lsm().

To avoid any concerns that IMA and EVM hooks are not called when migrating
from an old to a new kernel, always enable the 'integrity' LSM.

In addition, ensure that 'integrity' is the last LSM regardless of its
position in the list of supplied LSMs, as EVM needs to calculate the HMAC
after the other LSMs update their xattr.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>