security: Introduce file_pre_free_security hook
In preparation for moving IMA and EVM to the LSM infrastructure, introduce
the file_pre_free_security hook.

If the file was modified, IMA calculates at file close the new digest of
the file content and writes it to security.ima, so that appraisal at next
file access succeeds.

LSMs could also take some action while the file is still open before close.

The new hook cannot return an error and cannot cause the operation to be
reverted.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>
Reviewed-by: Stefan Berger <stefanb@linux.ibm.com>