KEYS: asymmetric: Introduce UMD-based asymmetric key parser
Introduce a new parser for the asymmetric key type. The parser forwards the
key payload to a User Mode Driver (UMD) handler, which replies with the
data necessary to populate a public_key structure (blob, IDs and algorithm)
and, possibly, with a key description.

The parser then validates the response, fills the key payload fields from
it, and sets the key description if provided.

The main advantage of this approach is that potential risks of parsing data
from the kernel are avoided, as the UMD handler does not have access to the
kernel memory.

If the UMD handler is corrupted, the damage is likely limited to user
space. The parser sanitizes malicious data sent by the UMD. In addition,
seccomp can further limit which system calls the UMD is authorized to
invoke.

The communication protocol between the parser and the UMD handler consists
of two main structures: msg_in for the parser request, and msg_out for the
UMD handler response.

msg_in contains a buffer for a generic request, not only for parsing keys,
and the type of command (CMD_KEY for key parsing).

msg_out contains the result of the operation to communicate to the parser,
and the specific structure umd_key_msg_out for storing the key payload
fields and the key description.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>