evm: Move to LSM infrastructure
As for IMA, move hardcoded EVM function calls from the LSM infrastructure
and the VFS to a new standalone LSM named 'evm'. Assign a new LSM ID to the
new LSM.

Make EVM functions as static (except for evm_inode_init_security(), which
is exported), and register them as hook implementations in init_evm_lsm(),
the new init method of the 'evm' LSM.

Finally, switch to the LSM reservation mechanism for the EVM xattr, and
consequently decrement by one the number of xattrs to allocate in
security_inode_init_security().

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>