integrity: Remove LSM
Since now IMA and EVM use their own integrity metadata, it is safe to
remove the integrity_iint_cache structure, and the associated functions
for its management.

Keep the iint.c file only for loading IMA and EVM keys at boot, and for
creating the integrity directory in securityfs (we need to keep it for
retrocompatibility reasons).

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>